// src/app/main/models/Invoice.ts
// Import the Order interface

import { Order } from "./Order";
import { Vendor } from "./Vendor";

export interface Invoice {
  id?: number; // Optional, auto-generated by the server
  invoiceNumber: string;
  invoiceDate: string; // Date formatted as YYYY-MM-DD
  vendor: Vendor; // Reference to the associated Vendor
  orders: Order[]; // List of associated Orders
  totalAmount: number;
  notes?: string; // Optional, additional notes
}

  